[Oztechan/CCC#1886] Move constants from infrastructure into model
