+++ backend/service/free/build.gradle.kts
@@ -21,7 +21,6 @@ kotlin {
                 }
                 Modules.Common.Core.apply {
                     implementation(project(network))
-                    implementation(project(infrastructure))
                     implementation(project(model))
                 }
             }
