+++ backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/controller/client/ClientControllerImpl.kt
@@ -3,9 +3,9 @@ package com.oztechan.ccc.backend.controller.client
 import co.touchlab.kermit.Logger
 import com.oztechan.ccc.backend.service.free.FreeApiService
 import com.oztechan.ccc.backend.service.premium.PremiumApiService
-import com.oztechan.ccc.common.core.infrastructure.constants.SECOND
 import com.oztechan.ccc.common.core.model.CurrencyType
 import com.oztechan.ccc.common.core.model.ExchangeRate
+import com.oztechan.ccc.common.core.model.constants.SECOND
 import com.oztechan.ccc.common.datasource.conversion.ConversionDataSource
 import kotlinx.coroutines.delay
