+++ client/repository/adcontrol/build.gradle.kts
@@ -23,7 +23,7 @@ kotlin {
                 implementation(project(Modules.Client.Storage.app))
                 implementation(project(Modules.Client.ConfigService.ad))
                 implementation(project(Modules.Client.Core.shared))
-                implementation(project(Modules.Common.Core.infrastructure))
+                implementation(project(Modules.Common.Core.model))
             }
         }
         val commonTest by getting {
