+++ client/src/commonTest/kotlin/com/oztechan/ccc/client/viewmodel/CurrenciesViewModelTest.kt
@@ -14,8 +14,8 @@ import com.oztechan.ccc.client.storage.app.AppStorage
 import com.oztechan.ccc.client.storage.calculator.CalculatorStorage
 import com.oztechan.ccc.client.viewmodel.currencies.CurrenciesEffect
 import com.oztechan.ccc.client.viewmodel.currencies.CurrenciesViewModel
-import com.oztechan.ccc.common.core.infrastructure.constants.SECOND
 import com.oztechan.ccc.common.core.model.Currency
+import com.oztechan.ccc.common.core.model.constants.SECOND
 import io.mockative.Mock
 import io.mockative.classOf
 import io.mockative.given
