+++ client/viewmodel/settings/build.gradle.kts
@@ -61,7 +61,6 @@ kotlin {
                     implementation(mockative)
                     implementation(coroutinesTest)
                 }
-                implementation(project(Modules.Common.Core.infrastructure))
             }
         }
