+++ client/viewmodel/watchers/build.gradle.kts
@@ -49,7 +49,6 @@ kotlin {
                     implementation(mockative)
                     implementation(coroutinesTest)
                 }
-                implementation(project(Modules.Common.Core.infrastructure))
             }
         }
