+++ client/viewmodel/widget/src/commonTest/kotlin/com/oztechan/ccc/client/viewmodel/widget/WidgetViewModelTest.kt
@@ -6,10 +6,10 @@ import com.oztechan.ccc.client.datasource.currency.CurrencyDataSource
 import com.oztechan.ccc.client.service.backend.BackendApiService
 import com.oztechan.ccc.client.storage.app.AppStorage
 import com.oztechan.ccc.client.storage.calculator.CalculatorStorage
-import com.oztechan.ccc.common.core.infrastructure.constants.DAY
 import com.oztechan.ccc.common.core.model.Conversion
 import com.oztechan.ccc.common.core.model.Currency
 import com.oztechan.ccc.common.core.model.ExchangeRate
+import com.oztechan.ccc.common.core.model.constants.DAY
 import io.mockative.Mock
 import io.mockative.classOf
 import io.mockative.given
