+++ common/core/model/src/commonMain/kotlin/com/oztechan/ccc/common/core/model/constants/TimeConstants.kt
@@ -1,4 +1,4 @@
-package com.oztechan.ccc.common.core.infrastructure.constants
+package com.oztechan.ccc.common.core.model.constants
 
 const val SECOND: Long = 1000L
 const val DAY: Long = 24 * 60 * 60 * SECOND
