+++ common/datasource/conversion/build.gradle.kts
@@ -28,7 +28,6 @@ kotlin {
                 }
                 Modules.Common.Core.apply {
                     implementation(project(database))
-                    implementation(project(infrastructure))
                     implementation(project(model))
                 }
             }
